+++ src/Kaponata.FileFormats/Lzma/LzmaCheck.cs
@@ -2,7 +2,7 @@
 // Copyright (c) Quamotion bv. All rights reserved.
 // </copyright>
 
-namespace Packaging.Targets.IO
+namespace Kaponata.FileFormats.Lzma
 {
     /// <summary>
     /// Type of the integrity check (Check ID).
