+++ src/Kaponata.FileFormats/Lzma/LzmaStream.cs
@@ -5,7 +5,7 @@
 using System;
 using System.Runtime.InteropServices;
 
-namespace Packaging.Targets.IO
+namespace Kaponata.FileFormats.Lzma
 {
     /// <summary>
     ///  Passing data to and from liblzma.
